Initial commit: setup project folder
Add virtualenv, gitignore, and README file. The virtualenv
folder is not being synced, so you'll have to install it manually
if you clone it. Later, if this scripts ends up using more libraries
than just BeautifulSoup, I will create a requirements.txt file
so you guys can install the dependencies easily.

Right now, the only requirement for this script is BeautifulSoup,
so I will not bother with requirements.txt.s